Wwise Audio Version 1.0
This one does not have errors, but I have still to implement sounds on the Lantern object, that is the one that was giving me errors.

Have Sounds on All interactable object to lvl2.
Ghost enemies are implemented correctly.

Only missing part for now is mixing and adding lantern Sound